Updating the project to use modules instead of having everything in our root dir
