+++ core/debug_windows_console.go
@@ -1,6 +1,6 @@
 //go:build windows && debug
 
-package main
+package core
 
 import (
 	"fmt"
