+++ core/platform_ext_windows.go
@@ -1,6 +1,6 @@
 //go:build windows
 
-package main
+package core
 
 import (
 	"encoding/json"
