+++ core/unix_stubs.go
@@ -1,6 +1,6 @@
 //go:build !windows
 
-package main
+package core
 
 import "os/exec"
