+++ gui/html/addgame.html
