+++ gui/html/style.css
