+++ gui/html/syncgame.js
