--- sync.go
@@ -1,5 +1,0 @@
-package main
-
-func SyncFilter(s string, syncDataPath Datapath) bool {
-	return false
-}
